Refactor tooltip to own component
