+++ packages/app/src/components-styled/time-series-chart/components/index.ts
@@ -2,3 +2,4 @@ export * from './chart-axes';
 export * from './chart-container';
 export * from './trend';
 export * from './markers';
+export * from './tooltip';
