+++ packages/app/src/components-styled/time-series-chart/logic/index.ts
@@ -1,1 +1,1 @@
-export * from './series-config';
+export * from './series';
